Login with database (and kotlin rewrite)
Added the login functionalities using the database. SharedPreferences is still used to keep the user logged in to the app if they close it, but actual user checking is done via queries. Rewrote the files to be in kotlin as leaving them as java would create issues with the database access (it also makes it easier to integrate). The foldername can be changed later to be more accurate, but when I tried to move the files all into components the app kept crashing so that would have to be debugged if organization is that important.