Login with database functionality
tables.kt is modified slightly from Henry's branch to include the Users table in the database. Logout and HomeScreen are modified due to me running into an issue with logging out and trying to log back in redirecting to the login screen. This was because the selectedscreen variable wasn't changed from the logout value so it got stuck in a loop. I added a small function to just reset it when Logout() is called to prevent this.