Fixing login issues
The Toast error messages being called in a thread caused the app to redirect to the home activity if a user entered different credentials after they successfully logged in and logged back out. This adds a variable to check instead and outputs Toast in the main thread to stop this from happening.